Fixing 15/06/2020

+++ repository/src/main/java/com/epam/esm/repository/certificate/tag/CertificateTagRepository.java
@@ -3,11 +3,11 @@ package com.epam.esm.repository.certificate.tag;
 import java.util.List;
 
 public interface CertificateTagRepository {
-    List<Long> getCertificateIdsByTagId(Long id);
+    List<Long> getCertificateIdsByTagId(Integer id);
 
-    List<Long> getTagIdsByCertificateId(Long id);
+    List<Integer> getTagIdsByCertificateId(Long id);
 
-    boolean saveCertificateTag(Long certId, Long tagId);
+    boolean saveCertificateTag(Long certId, Integer tagId);
 
-    boolean deleteCertificateTag(Long certId, Long tagId);
+    boolean deleteCertificateTag(Long certId, Integer tagId);
 }
